+++ lib/three3d/textures/DepthTexture.dart
@@ -1,7 +1,7 @@
 part of three_textures;
 
 class DepthTexture extends Texture {
-  DepthTexture(int width, int height, type, mapping, wrapS, wrapT, magFilter,
+  DepthTexture(double width, double height, type, mapping, wrapS, wrapT, magFilter,
       minFilter,
       anisotropy, format)
       : super(null, mapping, wrapS, wrapT, magFilter, minFilter, format, type,
